move non encoding methods from EncodingConfig to BroadcastConfig
