+++ CineIOExampleApp/src/main/java/io/cine/example/CineIoStreamViewActivity.java
@@ -13,10 +13,10 @@ import android.widget.TextView;
 import org.json.JSONException;
 import org.json.JSONObject;
 
+import io.cine.android.BroadcastConfig;
 import io.cine.android.CineIoClient;
 import io.cine.android.CineIoConfig;
 import io.cine.android.api.Stream;
-import io.cine.android.streaming.EncodingConfig;
 
 public class CineIoStreamViewActivity extends Activity {
 
@@ -61,7 +61,7 @@ public class CineIoStreamViewActivity extends Activity {
             @Override
             public void onClick(View view) {
                 Log.d(TAG, "Starting broadcast for " + stream.getId());
-                EncodingConfig config = new EncodingConfig();
+                BroadcastConfig config = new BroadcastConfig();
                 //TO SET A CUSTOM WIDTH AND HEIGHT
                 //config.setWidth(640);
                 //config.setHeight(480);
